add condition to migration as heroku complains

+++ db/migrate/20200826173739_delete_old_tables.rb
@@ -1,9 +1,9 @@
 class DeleteOldTables < ActiveRecord::Migration[6.0]
   def change
-    drop_table :exercises_rooms
-    drop_table :rooms_users
-    drop_table :rooms
-    drop_table :answers
-    drop_table :exercises
+    drop_table :exercises_rooms if ActiveRecord::Base.connection.table_exists?('exercises_rooms')
+    drop_table :rooms_users if ActiveRecord::Base.connection.table_exists?('rooms_users')
+    drop_table :rooms if ActiveRecord::Base.connection.table_exists?('rooms')
+    drop_table :answers if ActiveRecord::Base.connection.table_exists?('answers')
+    drop_table :exercises if ActiveRecord::Base.connection.table_exists?('exercises')
   end
 end
